// back-office-app/src/app/api/api-reference.ts
export enum TypeButton {
    Checkbox = 0,
    Radiobutton = 1,
}

export enum Category {
    Git = 0,
    Sql = 1,
    OOP = 2,
}
// ...
export class CreateOrUpdateQuestionCommand implements ICreateOrUpdateQuestionCommand {
    id?: number | undefined;
    category?: Category;
    typeButton?: TypeButton;
    text?: string;
    answers?: AnswerItem[];

    constructor(data?: ICreateOrUpdateQuestionCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.category = _data["category"];
            this.typeButton = _data["typeButton"];
            this.text = _data["text"];
            if (Array.isArray(_data["answers"])) {
                this.answers = [] as any;
                for (let item of _data["answers"])
                    this.answers!.push(AnswerItem.fromJS(item));
            }
        }
    }

    static fromJS(data: any): CreateOrUpdateQuestionCommand {
        data = typeof data === 'object' ? data : {};
        let result = new CreateOrUpdateQuestionCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["category"] = this.category;
        data["typeButton"] = this.typeButton;
        data["text"] = this.text;
        if (Array.isArray(this.answers)) {
            data["answers"] = [];
            for (let item of this.answers)
                data["answers"].push(item.toJSON());
        }
        return data;
    }
}

export interface ICreateOrUpdateQuestionCommand {
    id?: number | undefined;
    category?: Category;
    typeButton?: TypeButton;
    text?: string;
    answers?: AnswerItem[];
}

export class AnswerItem implements IAnswerItem {
    id?: number | undefined;
    isCorrect?: boolean;
    text?: string;

    constructor(data?: IAnswerItem) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.isCorrect = _data["isCorrect"];
            this.text = _data["text"];
        }
    }

    static fromJS(data: any): AnswerItem {
        data = typeof data === 'object' ? data : {};
        let result = new AnswerItem();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["isCorrect"] = this.isCorrect;
        data["text"] = this.text;
        return data;
    }
}

export interface IAnswerItem {
    id?: number | undefined;
    isCorrect?: boolean;
    text?: string;
}